+++ tests/__init__.py
@@ -6,9 +6,7 @@ from datetime import datetime, timedelta, timezone
 from flask.testing import FlaskClient
 from flask import url_for
 from flask_login import login_user
-from app.notify_client.models import (
-    InvitedOrgUser,
-)
+from app.models.user import InvitedOrgUser
 
 
 class TestClient(FlaskClient):
