+++ tests/app/main/views/organisations/test_organisation_invites.py
@@ -6,7 +6,7 @@ from bs4 import BeautifulSoup
 from flask import url_for
 from notifications_python_client.errors import HTTPError
 
-from app.notify_client.models import InvitedOrgUser
+from app.models.user import InvitedOrgUser
 from tests.conftest import ORGANISATION_ID, normalize_spaces
