+++ tests/app/main/views/test_accept_invite.py
@@ -6,7 +6,7 @@ from flask import url_for
 from notifications_python_client.errors import HTTPError
 
 import app
-from app.notify_client.models import InvitedUser
+from app.models.user import InvitedUser
 from tests.conftest import (
     SERVICE_ONE_ID,
     active_caseworking_user,
