+++ tests/app/main/views/test_activity.py
@@ -9,7 +9,7 @@ from flask import url_for
 from freezegun import freeze_time
 
 from app.main.views.jobs import get_status_filters, get_time_left
-from app.notify_client.models import Service
+from app.models.service import Service
 from tests.conftest import (
     SERVICE_ONE_ID,
     active_caseworking_user,
